# specs/005-workspace-completion/contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "EmptyInput": {
        "additionalProperties": false,
        "properties": {},
        "title": "EmptyInput",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ResourceResult": {
        "additionalProperties": false,
        "properties": {
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "revision": {
            "minimum": 0.0,
            "title": "Revision",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "revision"
        ],
        "title": "ResourceResult",
        "type": "object"
      },
      "ReviewAssistResult": {
        "additionalProperties": false,
        "properties": {
          "authorship_signal": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AuthorshipSignal"
              },
              {
                "type": "null"
              }
            ]
          },
          "feedback_draft": {
            "anyOf": [
              {
                "maxLength": 20000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Feedback Draft"
          },
          "suggestions": {
            "items": {
              "$ref": "#/components/schemas/ReviewerSuggestion"
            },
            "maxItems": 500,
            "minItems": 1,
            "title": "Suggestions",
            "type": "array"
          }
        },
        "required": [
          "suggestions"
        ],
        "title": "ReviewAssistResult",
        "type": "object"
      },
      "ReviewAssistView": {
        "additionalProperties": false,
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "result": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ReviewAssistResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "revision": {
            "minimum": 0.0,
            "title": "Revision",
            "type": "integer"
          },
          "status": {
            "enum": [
              "queued",
              "running",
              "unknown_outcome",
              "succeeded",
              "failed",
              "stale"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "status",
          "revision",
          "result",
          "error_code",
          "created_at"
        ],
        "title": "ReviewAssistView",
        "type": "object"
      },
...
      "ReviewDecision": {
        "additionalProperties": false,
        "properties": {
          "criterion_id": {
            "format": "uuid",
            "title": "Criterion Id",
            "type": "string"
          },
          "decision": {
            "enum": [
              "accepted",
              "changed",
              "manual"
            ],
            "title": "Decision",
            "type": "string"
          },
          "evidence_ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "maxItems": 100,
            "title": "Evidence Ids",
            "type": "array"
          },
          "points": {
            "minimum": 0.0,
            "title": "Points",
            "type": "number"
          },
          "reason": {
            "maxLength": 10000,
            "minLength": 1,
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "criterion_id",
          "points",
          "decision",
          "reason"
        ],
        "title": "ReviewDecision",
        "type": "object"
      },
      "ReviewDraftView": {
        "additionalProperties": false,
        "properties": {
          "criterion_decisions": {
            "items": {
              "$ref": "#/components/schemas/ReviewDecision"
            },
            "title": "Criterion Decisions",
            "type": "array"
          },
          "current_review_revision": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/WorkspaceRevisionSummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "current_review_revision_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Review Revision Id"
          },
          "review_notes": {
            "items": {
              "$ref": "#/components/schemas/WorkspaceNoteView"
            },
            "title": "Review Notes",
            "type": "array"
          },
          "revision": {
            "minimum": 0.0,
            "title": "Revision",
            "type": "integer"
          },
          "status": {
            "enum": [
              "queued",
              "in_review",
              "ready_to_publish",
              "published",
              "canceled"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "revision",
          "status",
          "current_review_revision_id",
          "current_review_revision",
          "criterion_decisions",
          "review_notes"
        ],
        "title": "ReviewDraftView",
        "type": "object"
      },
...
      "StudentHomeworkList": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/StudentHomeworkItem"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "minimum": 0.0,
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "minimum": 0.0,
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "offset",
          "limit"
        ],
        "title": "StudentHomeworkList",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "WorkList": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/WorkItem"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "minimum": 0.0,
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "minimum": 0.0,
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "offset",
          "limit"
        ],
        "title": "WorkList",
        "type": "object"
      },
...
      "WorkspaceCommand_EmptyInput_": {
        "additionalProperties": false,
        "properties": {
          "command_name": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Command Name",
            "type": "string"
          },
          "expected_revision": {
            "minimum": 0.0,
            "title": "Expected Revision",
            "type": "integer"
          },
          "idempotency_key": {
            "maxLength": 128,
            "minLength": 16,
            "title": "Idempotency Key",
            "type": "string"
          },
          "payload": {
            "$ref": "#/components/schemas/EmptyInput"
          },
          "request_id": {
            "format": "uuid",
            "title": "Request Id",
            "type": "string"
          },
          "target_id": {
            "format": "uuid",
            "title": "Target Id",
            "type": "string"
          }
        },
        "required": [
          "request_id",
          "idempotency_key",
          "command_name",
          "target_id",
          "expected_revision",
          "payload"
        ],
        "title": "WorkspaceCommand[EmptyInput]",
        "type": "object"
      },
...
      "WorkspaceCommand_WorkspaceReviewSaveInput_": {
        "additionalProperties": false,
        "properties": {
          "command_name": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Command Name",
            "type": "string"
          },
          "expected_revision": {
            "minimum": 0.0,
            "title": "Expected Revision",
            "type": "integer"
          },
          "idempotency_key": {
            "maxLength": 128,
            "minLength": 16,
            "title": "Idempotency Key",
            "type": "string"
          },
          "payload": {
            "$ref": "#/components/schemas/WorkspaceReviewSaveInput"
          },
          "request_id": {
            "format": "uuid",
            "title": "Request Id",
            "type": "string"
          },
          "target_id": {
            "format": "uuid",
            "title": "Target Id",
            "type": "string"
          }
        },
        "required": [
          "request_id",
          "idempotency_key",
          "command_name",
          "target_id",
          "expected_revision",
          "payload"
        ],
        "title": "WorkspaceCommand[WorkspaceReviewSaveInput]",
        "type": "object"
      },
      "WorkspaceNoteView": {
        "additionalProperties": false,
        "properties": {
          "author_user_id": {
            "format": "uuid",
            "title": "Author User Id",
            "type": "string"
          },
          "criterion_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Criterion Id"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "position": {
            "minimum": 0.0,
            "title": "Position",
            "type": "integer"
          },
          "text": {
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "id",
          "criterion_id",
          "text",
          "author_user_id",
          "position"
        ],
        "title": "WorkspaceNoteView",
        "type": "object"
      },
      "WorkspaceReviewSaveInput": {
        "additionalProperties": false,
        "properties": {
          "ai_run_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ai Run Id"
          },
          "draft": {
            "$ref": "#/components/schemas/SaveReviewRevisionPayload"
          },
          "signal_decisions": {
            "additionalProperties": {
              "enum": [
                "confirm",
                "reject"
              ],
              "type": "string"
            },
            "maxProperties": 1,
            "title": "Signal Decisions",
            "type": "object"
          }
        },
        "required": [
          "draft"
        ],
        "title": "WorkspaceReviewSaveInput",
        "type": "object"
      },
      "WorkspaceRevisionSummary": {
        "additionalProperties": false,
        "properties": {
          "author_user_id": {
            "format": "uuid",
            "title": "Author User Id",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "feedback": {
            "title": "Feedback",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "review_iteration_id": {
            "format": "uuid",
            "title": "Review Iteration Id",
            "type": "string"
          },
          "revision_number": {
            "minimum": 0.0,
            "title": "Revision Number",
            "type": "integer"
          },
          "total_score": {
            "title": "Total Score",
            "type": "number"
          }
        },
        "required": [
          "id",
          "review_iteration_id",
          "revision_number",
          "author_user_id",
          "feedback",
          "total_score",
          "created_at"
        ],
        "title": "WorkspaceRevisionSummary",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Review Workspace Extension",
    "version": "2.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v2/review-assists/{identity}": {
      "get": {
        "operationId": "get_assist_api_v2_review_assists__identity__get",
        "parameters": [
          {
            "in": "path",
            "name": "identity",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Identity",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewAssistView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Assist",
        "tags": [
          "workspace-v2"
        ]
      }
    },
    "/v2/review-assists/{identity}/retry": {
      "post": {
        "operationId": "retry_assist_api_v2_review_assists__identity__retry_post",
        "parameters": [
          {
            "in": "path",
            "name": "identity",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Identity",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkspaceCommand_EmptyInput_"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewAssistView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Retry Assist",
        "tags": [
          "workspace-v2"
        ],
        "x-command-name": "retry_review_assist"
      }
    },
...
    "/v2/reviews/{identity}/assist": {
      "get": {
        "operationId": "latest_assist_api_v2_reviews__identity__assist_get",
        "parameters": [
          {
            "in": "path",
            "name": "identity",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Identity",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/ReviewAssistView"
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "title": "Response Latest Assist Api V2 Reviews  Identity  Assist Get"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Latest Assist",
        "tags": [
          "workspace-v2"
        ]
      },
      "post": {
        "operationId": "start_assist_api_v2_reviews__identity__assist_post",
        "parameters": [
          {
            "in": "path",
            "name": "identity",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Identity",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkspaceCommand_EmptyInput_"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewAssistView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Start Assist",
        "tags": [
          "workspace-v2"
        ],
        "x-command-name": "start_review_assist"
      }
    },
...
    "/v2/reviews/{identity}/draft": {
      "get": {
        "operationId": "get_review_draft_api_v2_reviews__identity__draft_get",
        "parameters": [
          {
            "in": "path",
            "name": "identity",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Identity",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewDraftView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Review Draft",
        "tags": [
          "workspace-v2"
        ]
      },
      "post": {
        "operationId": "save_assisted_draft_api_v2_reviews__identity__draft_post",
        "parameters": [
          {
            "in": "path",
            "name": "identity",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Identity",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkspaceCommand_WorkspaceReviewSaveInput_"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResourceResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Save Assisted Draft",
        "tags": [
          "workspace-v2"
        ],
        "x-command-name": "save_workspace_review"
      }
    },
...
    "/v2/student/homeworks": {
      "get": {
        "operationId": "list_student_homeworks_api_v2_student_homeworks_get",
        "parameters": [
          {
            "in": "query",
            "name": "state",
            "required": false,
            "schema": {
              "default": "",
              "enum": [
                "",
                "in_progress",
                "completed"
              ],
              "title": "State",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 30,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentHomeworkList"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Student Homeworks",
        "tags": [
          "workspace-v2"
        ]
      }
    },
...
    "/v2/works": {
      "get": {
        "operationId": "works_api_v2_works_get",
        "parameters": [
          {
            "in": "query",
            "name": "course_run_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Course Run Id"
            }
          },
          {
            "in": "query",
            "name": "homework_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Homework Id"
            }
          },
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "default": "",
              "maxLength": 200,
              "title": "Q",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "state",
            "required": false,
            "schema": {
              "default": "",
              "title": "State",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "view",
            "required": false,
            "schema": {
              "default": "all",
              "title": "View",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "priority",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "assigned",
                    "deadline"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Priority"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 30,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkList"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Works",
        "tags": [
          "workspace-v2"
        ]
      }
    }
  },
  "servers": [
    {
      "url": "/api"
    }
  ]
}